pdf2txt/requirements.txt
Flask==2.3.2
google-cloud-storage==2.10.0
pdfplumber==0.9.0